% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-functions.R
\name{as_tidygraph}
\alias{as_tidygraph}
\title{Convert a tree to a \code{tidygraph} object.}
\usage{
as_tidygraph(tree)
}
\arguments{
\item{tree}{A list of tibbles, each corresponding to a generation
in the infection propagation}
}
\value{
A \code{tidygraph} object.
}
\description{
Convert a tree to a \code{tidygraph} object.
}
