% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{average_component_size}
\alias{average_component_size}
\title{The average component size of paths that go extinct.}
\usage{
average_component_size(u, A = 10, B = 1, Lambda = 0.11, P = 0.5)
}
\arguments{
\item{u}{The extinction probability}

\item{A}{The shape parameter of the gamma life time distribution. Default a =10}

\item{B}{The rate parameter of the gamma life time distribution. Default b = 1}

\item{Lambda}{The arrival rate of infectious interactions. Default lambda = .11}

\item{P}{The parameter of the logarithmic distribution for the number of infected during an event.
Default p=0.5}
}
\value{
The average component size.
}
\description{
The average component size of paths that go extinct.
}
