% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchsim-package.R
\docType{package}
\name{branchsim-package}
\alias{branchsim}
\alias{branchsim-package}
\title{branchsim: A branching process simulation of communicable disease}
\description{
Simulates the evolution of a communicable disease as a branching process. The model allows for the stochastic interruption of a communicable period, to model the effect of contact tracing, for example.
}
\author{
\strong{Maintainer}: Jérôme Levesque \email{jerome.levesque@tpsgc-pwgsc.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
