% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{calc_avg_path}
\alias{calc_avg_path}
\title{Compute average epidemic path}
\usage{
calc_avg_path(
  lambda = 0.11,
  mu = 1.5,
  A = 5.5,
  B = 0.85,
  tmax = 100,
  nstep = 10000
)
}
\description{
Compute average epidemic path
}
