% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{calc_mu}
\alias{calc_mu}
\title{Expectation of the logarithmic distribution}
\usage{
calc_mu(p)
}
\arguments{
\item{p}{A scalar greater than 0 and smaller than 1.}
}
\value{
The expectation value of the logarithmic distribution.
}
\description{
Expectation of the logarithmic distribution
}
