% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{calc_R0}
\alias{calc_R0}
\title{Calculate R0 implied from the model}
\usage{
calc_R0(tbar, lambda, mu, q = 0, mbar = NA)
}
\arguments{
\item{tbar}{Average duration of the communicable window.}

\item{lambda}{Poisson rate}

\item{mu}{Expectation of the logarithmic distribution giving the
number of individuals infected per event}

\item{q}{Probability of interruption}

\item{mbar}{Average communicable period elapsed at the time of
interuption}
}
\value{
The value of \eqn{R_0}.
}
\description{
Calculate R0 implied from the model
}
