% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{char_function}
\alias{char_function}
\title{The characteristic function of the branching process.}
\usage{
char_function(u, a = 10, b = 1, lambda = 0.11, p = 0.5)
}
\arguments{
\item{u}{The argument of the geberating function.}

\item{a}{The shape parameter of the gamma life time distribution. Default a =10}

\item{b}{The rate parameter of the gamma life time distribution. Default b = 1}

\item{lambda}{The arrival rate of infectious interactions. Default lambda = .11}

\item{p}{The parameter of the logarithmic distribution for the number of infected during an event.
Default p=0.5}
}
\value{
The evaluation the characteristic function at u.
}
\description{
The characteristic function of the branching process.
}
