% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{dMu}
\alias{dMu}
\title{A function for the infinitesimal expectation in births: mu(dt) = dMu(t) dt}
\usage{
dMu(t, a = 10, b = 1, lambda = 0.11, p = 0.5)
}
\arguments{
\item{t}{A vector of time evaluations}

\item{a}{The shape parameter of the gamma life time distribution. Default a =10}

\item{b}{The rate parameter of the gamma life time distribution. Default b = 1}

\item{lambda}{The arrival rate of infectious interactions. Default lambda = .11}

\item{p}{The parameter of the logarithmic distribution for the number of infected during an event.
Default p=0.5}
}
\value{
Function evaluation at time vector t.
}
\description{
A function for the infinitesimal expectation in births: mu(dt) = dMu(t) dt
}
