% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-functions.R
\name{equalize_paths}
\alias{equalize_paths}
\title{Make all paths end at a fixed time}
\usage{
equalize_paths(df_paths, tmax)
}
\arguments{
\item{df_paths}{A \code{data.frame}. Typically, the output of
\code{\link{treelist_to_paths}}. The following columns are
required: \code{date}, \code{id_sim}.}

\item{tmax}{A non-negative scalar.}
}
\value{
A \code{tbl} with the same format as \code{df_paths}, but
with the last point for each path fixed at \code{tmax}.
}
\description{
Make all paths end at a fixed time
}
