% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-plotting-functions.R
\name{extinct_plot}
\alias{extinct_plot}
\title{A gradient plot of the extincition probability as a function of the average number of infectious events per day
and average number infected per infectious event.}
\usage{
extinct_plot(lambda_limits, mu_limits, a = 5.5, b = 0.85)
}
\arguments{
\item{lambda_limits}{A vector of the lower and upper bounds of the average arrival rate of interest.}

\item{mu_limits}{A vector of the lower and upper bounds of the average number infected per interaction.}

\item{a}{The shape parameter of the communicable preiod gamma function. Default a =5.5.}

\item{b}{The rate parameter of the communicable preiod gamma function. Default b = 0.85.}
}
\value{
A plot showing a gradient of the extinction probability.
}
\description{
A gradient plot of the extincition probability as a function of the average number of infectious events per day
and average number infected per infectious event.
}
