% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{find_p}
\alias{find_p}
\title{A function for obtaining the p parameter of the log-series distribution from its mean, mu.}
\usage{
find_p(mu)
}
\arguments{
\item{mu}{The mean of a log-series distribution.}
}
\value{
The p value of the log-series distribution
}
\description{
A function for obtaining the p parameter of the log-series distribution from its mean, mu.
}
