% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{G}
\alias{G}
\title{A function for the expecation for the number of idividuals existing in the process at moment t.
Used with the renewal eqaution, it will give the expectation for the instantaneous number infected
as a function of time.}
\usage{
G(t, a = 10, b = 1)
}
\arguments{
\item{t}{A vector of time evaluations}

\item{a}{The shape parameter of the gamma life time distribution. Default a =10}

\item{b}{The rate parameter of the gamma life time distribution. Default b = 1}
}
\value{
Function evaluation at time vector t.
}
\description{
A function for the expecation for the number of idividuals existing in the process at moment t.
Used with the renewal eqaution, it will give the expectation for the instantaneous number infected
as a function of time.
}
