% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{generating_function}
\alias{generating_function}
\title{The generating function of the branching process f(s) = sum_k p_k s^k.}
\usage{
generating_function(s, a = 10, b = 1, lambda = 0.11, p = 0.5)
}
\arguments{
\item{s}{The argument of the generating function. 0<s<1.}

\item{a}{The shape parameter of the gamma life time distribution. Default a =10}

\item{b}{The rate parameter of the gamma life time distribution. Default b = 1}

\item{lambda}{The arrival rate of infectious interactions. Default lambda = .11}

\item{p}{The parameter of the logarithmic distribution for the number of infected during an event.
Default p=0.5}
}
\value{
The evaluation the generating function at s.
}
\description{
The generating function of the branching process f(s) = sum_k p_k s^k.
}
