% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{make_first_layer}
\alias{make_first_layer}
\title{Initialize the tree}
\usage{
make_first_layer(n_start, tbar, kappa)
}
\arguments{
\item{n_start}{An integer greater than or equal to 1. Number of
nodes in the initial layer.}

\item{tbar}{A non-negative scalar. Average duration of the
communicable windows of the initial nodes. Together with
\code{kappa}, it determines a shape parameter \code{tbar *
    kappa} for the the Gamma distribution from which we draw a
communicable period for each new node.}

\item{kappa}{A non-negative scalar (including \code{Inf}). Rate
parameter for the Gamma distribution from which we draw a
communicable period duration for each new node. If \code{kappa
    == Inf}, then the draw is deterministic, with all communicable
periods exactly equal to \code{tbar}.}
}
\value{
A \code{tibble} with one row for each new node, and five
columns: \code{id}, \code{id_parent}, \code{id_layer},
\code{t_infect}, and \code{t_comm}.
}
\description{
Creates the first layer of an infection tree.
}
