% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot dendrogram corresponding to a single path}
\usage{
plot_dendrogram(single_path_data, tmax = NA)
}
\description{
Plot dendrogram corresponding to a single path
}
