% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_exp}
\alias{plot_exp}
\title{Plot the time between infectious events}
\usage{
plot_exp(lambda, pmax = 0.995, n = 1000)
}
\arguments{
\item{lambda}{Poisson rate}

\item{pmax}{A scalar between 0 and 1. Upper limit for the time axis.}

\item{n}{Number of points to plot.}
}
\value{
A ggplot object
}
\description{
Plot the time between infectious events
}
