% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot the Gamma density for the communicable window}
\usage{
plot_gamma(tbar, kappa, pmax = 0.999, n = 1000)
}
\arguments{
\item{tbar}{Average duration of the communicable window.}

\item{kappa}{Rate parameter of the Gamma distribution.}

\item{pmax}{A scalar between 0 and 1. Upper limit for the time axis.}

\item{n}{Number of points to plot.}
}
\value{
A ggplot object
}
\description{
Plot the Gamma density for the communicable window
}
