% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-plotting-functions.R
\name{plot_gamma_dist}
\alias{plot_gamma_dist}
\title{The plot of communicable window's distribution.}
\usage{
plot_gamma_dist(a = 5.5, b = 0.85)
}
\arguments{
\item{a}{The shape parameter of the communicable preiod gamma function.}

\item{b}{The rate parameter of the communicable preiod gamma function.}
}
\value{
A plot showing a the distribution function of the communicable period.
}
\description{
The plot of communicable window's distribution.
}
