% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_log}
\alias{plot_log}
\title{Plot the number of individuals infected per infectious event}
\usage{
plot_log(p, q_high = 0.995)
}
\arguments{
\item{p}{Parameter of the logarithmic distribution.}

\item{q_high}{Upper quantile to bound the plot.}
}
\value{
A ggplot object
}
\description{
Plot the number of individuals infected per infectious event
}
