% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot epidemic paths}
\usage{
plot_paths(paths_data, y_var = n_infected, n_max = 200)
}
\arguments{
\item{paths_data}{A \code{data.frame}, with at least the columns
\code{id_sim}, \code{date}, and a quoted variable passed as
\code{yvar}.}

\item{y_var}{Column in \code{paths_data} to use as y-variable.}

\item{n_max}{An integer. Maximum number of paths to display. Useful
to avoid graphs that are too crowded.}
}
\description{
Plot epidemic paths
}
