% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-plotting-functions.R
\name{plot_single_mother_dist}
\alias{plot_single_mother_dist}
\title{The plot of a single mother's distribution function for her infected population.}
\usage{
plot_single_mother_dist(a = 5.5, b = 0.85, lambda = 0.11, mu = 1.5)
}
\arguments{
\item{a}{The shape parameter of the communicable preiod gamma function.}

\item{b}{The rate parameter of the communicable preiod gamma function.}

\item{lambda}{The arrival rate of infectious interactions.}

\item{mu}{The average number infected per infectious interaction}
}
\value{
A plot showing a the distribution function of a single mother's infected population
}
\description{
The plot of a single mother's distribution function for her infected population.
}
