% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_single_path}
\alias{plot_single_path}
\title{Show infection counts (cumulative and ongoing) for a single path}
\usage{
plot_single_path(
  path_data,
  tlim = NA,
  ylim = NA,
  scale = "log",
  smooth_data = NULL
)
}
\arguments{
\item{path_data}{A \code{data.frame}, with at least the columns
\code{time}, \code{n_infected}, and \code{n_infectious}}

\item{tlim}{Upper time limit to display. If \code{NA}, show the
entire path.}

\item{ylim}{Explicit limit for the top of the graph. Useful when
showing the detail of one path in comparison with the output of
\code{plot_paths}, so that the y-scales are the same between
the two plots.}

\item{scale}{If equal to "log", show in log scale (default). If any
other value, show in linear scale. Like argument \code{ylim},
this is useful when comparing to the output of
\code{plot_paths}.}

\item{smooth_data}{A \code{data.frame}, with at least the columns
\code{time}, \code{n_infected}, and \code{n_infectious}. If
given, will be superposed on top of the paths plot.}
}
\value{
A \code{ggplot} plot showing the cumulative infection
count, the ongoing infection count, and, if \code{smooth_data}
is given, the expectation values of the stochastic process.
}
\description{
Show infection counts (cumulative and ongoing) for a single path
}
