% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{prob_ninfect}
\alias{prob_ninfect}
\title{The probability distribution of births from a single mother in the branching process.}
\usage{
<<<<<<< HEAD:man/prob_distribution.Rd
prob_distribution(tbar, kappa, lambda, p, n_samp = 300000L, min_count = 4)
=======
prob_ninfect(tbar, kappa, lambda, p, n_samp = 300000L, min_count = 4)
>>>>>>> 56a8e3c3d15f93fe0dbe545dc10901eb6edd98e8:man/prob_ninfect.Rd
}
\arguments{
\item{lambda}{The arrival rate of infectious interactions. Default lambda = .11}

\item{p}{The parameter of the logarithmic distribution for the number of infected during an event.
Default p=0.5}

\item{a}{The shape parameter of the gamma life time distribution. Default a =10}

\item{b}{The rate parameter of the gamma life time distribution. Default b = 1}
}
\value{
A tibble of counts with probability.
}
\description{
The probability distribution of births from a single mother in the branching process.
}
