% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{renewal_function}
\alias{renewal_function}
\title{Renewal equation solution to the expectation of the Crump-Mode-Jagers (CMJ) over a random characteristic.}
\usage{
renewal_function(dMu, G = 1, T = 100, nstep = 10000)
}
\arguments{
\item{dMu}{The function dMu(t) that defines mu(dt) = dMu(t) dt.}

\item{G}{A function that that defines the expectation of the random characteristic of interest. Set to G=1
by default (total number born).}

\item{T}{The maximum of the time interval for integration 0,T.}

\item{nstep}{The number of steps for the integration.}
}
\value{
A tibble containing the time steps with the solution to the renewal equation.
}
\description{
This function numerically solves the renewal equation for a CMJ process using the properties of the triple
(lambda_chi, xi_chi, chi). See appendix C of Branching Processes in Biology by Kimmel and Axelrod for more details.
Z = G(t) + int_0^t Z(t - u) mu(du)
}
