% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-plotting-functions.R
\name{renewal_plot}
\alias{renewal_plot}
\title{Plot the solution to the renewal equation solution for expectations of the Crump-Mode-Jagers (CMJ) over a random characteristic.}
\usage{
renewal_plot(
  Lambda = 0.11,
  mu = 1.5,
  A = 5.5,
  B = 0.85,
  Time = 100,
  type = "Total"
)
}
\arguments{
\item{Lambda}{The arrival rate of infectious interactions.}

\item{mu}{The average number infected per infectious interaction}

\item{A}{The shape parameter of the communicable preiod gamma function.}

\item{B}{The rate parameter of the communicable preiod gamma function.}

\item{Time}{The end of the time interval of interest \code{[0,T]}.}

\item{type}{The type or random characteristic Total for total infected Infectious for currently infected}
}
\value{
A plot showing the solution of the renewal equation with the malthusian parameter.
}
\description{
This function plots the numerical soution to the renewal equation for a CMJ process using the properties of the triple
(lambda_chi, xi_chi, chi). See appendix C of Branching Processes in Biology by Kimmel and Axelrod for more details.
Z = G(t) + int_0^t Z(t - u) mu(du)
}
