% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{run_sims}
\alias{run_sims}
\title{Multiple path simulations over a range of input parameters}
\usage{
run_sims(
  nsim = 10,
  nstart = 1,
  tmax = 100,
  tbar = 10,
  p = 0.5,
  lambda = 0.11,
  kappa = 1,
  q = 0.6,
  mbar = 5,
  kappaq = 3,
  ceil = Inf
)
}
\arguments{
\item{nsim}{A non-negative integer. The number of replications for
each set of input parameters.}

\item{...}{Input parameters for \code{\link{add_layer}}. Each input
parameter can receive a vector with multiple values. If that is
the case, then all input parameters should have the same
length, or one.}
}
\value{
A tibble with a column for each input parameter, and a
list-column \code{treelist}. This is a list of lists of lists.
Each sublist corresponds to a unique combination of input
parameters, and contains \code{nsim} trees (i.e. simulated
paths). Each individual tree is itself a list of tibbles, with
one tibble for each depth level.
}
\description{
This function calls \code{\link{build_tree}} over a range of input
parameters. It computes \code{nsim} replications for each unique
combination of input parameters.
}
