% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{sim_ninfections}
\alias{sim_ninfections}
\title{Simulate the distribution of individuals infected per patient}
\usage{
sim_ninfections(tbar, kappa, lambda, p, n = 1e+05)
}
\arguments{
\item{tbar}{Average duration of the communicable window.}

\item{kappa}{Rate parameter of the Gamma distribution.}

\item{lambda}{Poisson rate}

\item{p}{Logarithmic distribution parameter}

\item{n}{Number of points to simulate}
}
\value{
A vector of integers of length \code{n}, giving the number of
individuals infected by each infectious individual
}
\description{
Simulate the distribution of individuals infected per patient
}
