% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchsim-package.R
\docType{package}
\name{branchsim-package}
\alias{branchsim}
\alias{branchsim-package}
\title{branchsim: Branching process simulation of epidemics}
\description{
Simulates the propagation of an epidemic as a branching process. The model allows for the stochastic interruption of communicable periods to study the effect of contact tracing, for example.
}
\author{
\strong{Maintainer}: Jérôme Levesque \email{jerome.levesque@tpsgc-pwgsc.gc.ca}

Authors:
\itemize{
  \item David Maybury \email{david.maybury@tpsgc-pwgsc.gc.ca}
}

}
\keyword{internal}
