% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{dMu}
\alias{dMu}
\title{A function for the infinitesimal expectation in births: mu(dt) = dMu(t) dt}
\usage{
dMu(A = 10, B = 1, Lambda = 0.11, P = 0.5)
}
\arguments{
\item{A}{The shape parameter of the gamma life time distribution. Default A =10}

\item{B}{The rate parameter of the gamma life time distribution. Default B = 1}

\item{Lambda}{The arrival rate of infectious interactions. Default Lambda = .11}

\item{P}{The parameter of the logarithmic distribution for the number of infected during an event.
Default P=0.5}
}
\value{
A function of t: dMu(t).
}
\description{
A function for the infinitesimal expectation in births: mu(dt) = dMu(t) dt
}
