% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{find_kappa}
\alias{find_kappa}
\title{Find Gamma rate parameter}
\usage{
find_kappa(tbar, p_val, p = 0.95)
}
\arguments{
\item{q}{}
}
\value{
A shape parameter higher than zero
}
\description{
Find Gamma rate parameter
}
