% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-functions.R
\name{G}
\alias{G}
\title{A function for the expecation for the number of individuals existing in the process at moment t.
Used with the renewal equation, it will give the expectation for the instantaneous number infected
as a function of time.}
\usage{
G(A = 10, B = 1)
}
\arguments{
\item{A}{The shape parameter of the gamma life time distribution. Default A =10}

\item{B}{The rate parameter of the gamma life time distribution. Default B = 1}
}
\value{
A function \eqn{G(t) = 1 - F(t)}; F(t) is the life-time distribution.
}
\description{
A function for the expecation for the number of individuals existing in the process at moment t.
Used with the renewal equation, it will give the expectation for the instantaneous number infected
as a function of time.
}
