% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-functions.R
\name{make_path}
\alias{make_path}
\title{Leaving this function here for now, but will probably remove soon.
\code{\link{treelist_to_paths}} is much faster.}
\usage{
make_path(tree_raw, tmax = Inf)
}
\description{
Leaving this function here for now, but will probably remove soon.
\code{\link{treelist_to_paths}} is much faster.
}
