% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_log}
\alias{plot_log}
\title{Plot the number of individuals infected per infectious event}
\usage{
plot_log(p, q_high = 0.995)
}
\arguments{
\item{mu}{Average number of individuals infected}
}
\value{
A ggplot object
}
\description{
Plot the number of individuals infected per infectious event
}
