% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot epidemic paths}
\usage{
plot_paths(
  paths_data,
  y_var = n_infected,
  n_max = 200,
  tlim = NA,
  highlight_path_id = NA,
  show_hist = TRUE,
  show_extinct = TRUE,
  smooth_data = NULL,
  extinct_color = "red",
  n_bins = 100
)
}
\arguments{
\item{paths_data}{A \code{data.frame}, with at least the columns
\code{id_sim}, \code{time}, and a quoted variable passed as
\code{yvar}.}

\item{y_var}{Column in \code{paths_data} to use as y-variable.}

\item{n_max}{An integer. Maximum number of paths to display. Useful
to avoid graphs that are too crowded.}

\item{tlim}{A scalar.}

\item{show_hist}{A logical scalar. If \code{TRUE}, show the
distribution of paths at the end.}

\item{show_extinct}{A logical scalar. Show extinct paths along with
non-extinct ones.}

\item{smooth_data}{A \code{data.frame}, with at least the columns
\code{time}, \code{n_infected}, and \code{n_infectious}.}

\item{extinct_color}{A color value understood by ggplot.}

\item{n_bins}{An integer scalar. Number of bins in the histogram
panel.}
}
\value{
A plot.
}
\description{
Plot epidemic paths
}
