% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot epidemic paths}
\usage{
plot_paths(
  paths_data,
  y_var = n_infected,
  n_max = 200,
  show_hist = TRUE,
  show_1_mass = FALSE
)
}
\arguments{
\item{paths_data}{A \code{data.frame}, with at least the columns
\code{id_sim}, \code{time}, and a quoted variable passed as
\code{yvar}.}

\item{y_var}{Column in \code{paths_data} to use as y-variable.}

\item{n_max}{An integer. Maximum number of paths to display. Useful
to avoid graphs that are too crowded.}

\item{show_hist}{A logical scalar. If \code{TRUE}, show the
distribution of paths at the end.}

\item{show_1_mass}{A logical scalar. Show the probability mass at
\code{y_var = 1}. When \code{y_var} is \code{n_infected}, this
corresponds to epidemics that did not continue past the first
infected individual. NOTE: this parameter is meaningful only
when \code{show_hist} is \code{TRUE}.}
}
\value{
A plot.
}
\description{
Plot epidemic paths
}
