% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_R0}
\alias{plot_R0}
\title{Contour plot of \eqn{R_0} for different values of \eqn{\lambda} and \eqn{\mu}.}
\usage{
plot_R0(
  my_lambda,
  my_mu,
  tbar,
  lambda_lim = c(0.01, 0.2),
  mu_lim = c(1.01, 2),
  q = 0,
  mbar = NA,
  n = 32
)
}
\arguments{
\item{lambda}{A vector of two values covering the range of \eqn{lambda}.}

\item{mu}{A vector of two values covering the range of \eqn{mu}.}
}
\value{
A ggplot object
}
\description{
Contour plot of \eqn{R_0} for different values of \eqn{\lambda} and \eqn{\mu}.
}
