% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMJ-plotting-functions.R
\name{renewal_plot}
\alias{renewal_plot}
\title{Plot the solution to the renewal equation solution for expecations of the Crump-Mode-Jagers (CMJ) over a random characteristic.}
\usage{
renewal_plot(
  Lambda = 0.11,
  mu = 1.5,
  A = 5.5,
  B = 0.85,
  Time = 100,
  type = "Total"
)
}
\arguments{
\item{Lambda}{The function dMu(t) that defines mu(dt) = dMu(t) dt. Function as argument.}

\item{mu}{A function that that defines the expectation of the random characteristic of interest. Set to G=1
by default (total number born). Function as argument.}

\item{A}{The maximum of the time interval for integration 0,T.}

\item{B}{The number of steps for the integration.}

\item{Time}{The number of steps for the integration.}

\item{type}{The number of steps for the integration.}
}
\value{
A tibble containing the time steps with the solution to the renewal equation.
}
\description{
This function plots the numerical soution to the renewal equation for a CMJ process using the properties of the triple
(lambda_chi, xi_chi, chi). See appendix C of Branching Processes in Biology by Kimmel and Axelrod for more details.
Z = G(t) + int_0^t Z(t - u) mu(du)
}
